refactor(orchestration): extract tool execution pipeline and messages builder
agent-loop.ts の executeToolCall クロージャ（180行+）と buildMessagesForAPI
など履歴組み立てロジックを別モジュールに切り出し、エージェントループ本体は
「ターンを回す」責務に近づける。

- `tool-execution-pipeline.ts`: ツール実行 + 例外→Result 包装 + security
  middleware + screenshot 置換 + bg_fetch SPA 警告追記を集約。repl 用の
  console-log hook 生成も同居。
- `messages-builder.ts`: `buildMessagesForAPI` / `toPersistedHistory` /
  `normalizeImageForApi` / `buildAssistantMessageContent` と、session summary
  / 構造化要約 / legacy tool_result マーカーの除去ヘルパ群を集約。
- `visited-url-tracker.ts`: ツール結果起点の URL 追跡 (`trackVisitedUrlsFromToolResult`)
  と navigate/repl 直後のタブ URL 確認 (`detectUrlChangeAfterNavTool`) を追加。
- agent-loop.ts は上記を呼ぶだけに縮小。executeToolCall は 180+ 行 → 86 行。
  nav フラグ / security 通知は turn スコープの小ヘルパに集約。
- `normalizeImageForApi` / `toPersistedHistory` は既存テスト互換のため
  re-export を維持。

UI state への直接アクセス（`useStore.getState()`）は本コミットではまだ残して
いる（artifact 自動展開 / session summary sync / enableSecurityMiddleware）。
それは Step 3 で callback 注入に置き換える。

Refs: #110
